+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-23 02:56:55
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: dedc7d50a4615b5b6fd1aa992e91fae5d6346a7e
+
+## [ForwardWidgets-2kuai] - 2025-10-23 02:56:55
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 71383d350cd1879102607564fcf57fcc7a04e1a1
+
+## [ForwardWidgets-pack1r] - 2025-10-23 02:56:55
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 1c8c381d46b4869ab12cca488dbd4fd7d2f4fa8f
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-22 03:03:20
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main